src/KeyVaultFunctions/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform
azure-functions
azure.core
azure-keyvault-certificates
azure-identity==1.11.0